requirements.txt
nltk==3.6.5
numpy==1.23.3
pandas==1.3.4
PyYAML==6.0
scikit_learn==1.1.1
tensorflow==2.9.1